Move ConfigurationParser interface into common package to reduce package cycles.
