+++ project-set/commons/classloader/src/main/java/com/rackspace/papi/commons/util/classloader/ear/DefaultEarArchiveEntryListener.java
@@ -4,7 +4,7 @@ import com.oracle.javaee6.ApplicationType;
 import com.oracle.javaee6.FilterType;
 import com.oracle.javaee6.ObjectFactory;
 import com.oracle.javaee6.WebFragmentType;
-import com.rackspace.papi.commons.config.parser.ConfigurationParser;
+import com.rackspace.papi.commons.config.parser.common.ConfigurationParser;
 import com.rackspace.papi.commons.config.parser.jaxb.JaxbConfigurationParser;
 import com.rackspace.papi.commons.config.resource.impl.ByteArrayConfigurationResource;
 import com.rackspace.papi.commons.util.StringUtilities;
@@ -18,7 +18,6 @@ import com.rackspace.papi.commons.util.plugin.archive.ProcessingAction;
 import java.io.File;
 import java.util.Arrays;
 import java.util.List;
-import java.util.jar.Attributes;
 import java.util.jar.Manifest;
 import javax.xml.bind.JAXBContext;
 import javax.xml.bind.JAXBElement;
