+++ project-set/commons/configuration/src/main/java/com/rackspace/papi/commons/config/manager/ConfigurationUpdateManager.java
@@ -1,7 +1,7 @@
 package com.rackspace.papi.commons.config.manager;
 
 import com.rackspace.papi.commons.config.resource.ConfigurationResource;
-import com.rackspace.papi.commons.config.parser.ConfigurationParser;
+import com.rackspace.papi.commons.config.parser.common.ConfigurationParser;
 import com.rackspace.papi.commons.util.Destroyable;
 
 public interface ConfigurationUpdateManager extends Destroyable {
