+++ project-set/commons/configuration/src/main/java/com/rackspace/papi/commons/config/parser/common/AbstractConfigurationObjectParser.java
@@ -1,7 +1,5 @@
 package com.rackspace.papi.commons.config.parser.common;
 
-import com.rackspace.papi.commons.config.parser.ConfigurationParser;
-
 public abstract class AbstractConfigurationObjectParser<T> implements ConfigurationParser<T> {
 
     private final Class<T> configurationClass;
