+++ project-set/commons/configuration/src/main/java/com/rackspace/papi/commons/config/parser/common/ConfigurationParser.java
@@ -1,4 +1,4 @@
-package com.rackspace.papi.commons.config.parser;
+package com.rackspace.papi.commons.config.parser.common;
 
 import com.rackspace.papi.commons.config.resource.ConfigurationResource;
