+++ project-set/core/core-lib/src/main/java/com/rackspace/papi/service/config/ParserListenerPair.java
@@ -1,7 +1,7 @@
 package com.rackspace.papi.service.config;
 
 import com.rackspace.papi.commons.config.manager.UpdateListener;
-import com.rackspace.papi.commons.config.parser.ConfigurationParser;
+import com.rackspace.papi.commons.config.parser.common.ConfigurationParser;
 
 import java.lang.ref.WeakReference;
